third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-2/keyframes-rules-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that source data is extracted correctly from stylesheets with @keyframes rules.


Running: testInit

Running: testDumpStyles
[expanded] 
element.style { ()
    background-color: white;

[expanded] 
* { (<style>…</style>)
/-- overloaded --/     background-color: papayawhip;

[expanded] 
div { (user agent stylesheet)
    display: block;
